requirements.txt
# 深度学习框架
torch>=2.0.0
torchvision>=0.15.0

# 图像处理与计算机视觉
opencv-python>=4.8.0
Pillow>=9.0.0

# 数据处理与科学计算
numpy>=1.24.0
pandas>=2.0.0

# 数据可视化与绘图
matplotlib>=3.7.0
seaborn>=0.12.0

# 机器学习评估指标 (混淆矩阵等)
scikit-learn>=1.2.0

# 进度条工具
tqdm>=4.65.0

# 训练日志记录
tensorboard>=2.13.0